requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
bidict=0.21.2=pyhd3eb1b0_0
bzip2=1.0.8=h80987f9_5
ca-certificates=2024.3.11=hca03da5_0
click=8.1.7=py312hca03da5_0
expat=2.5.0=h313beb8_0
flask=2.2.5=py312hca03da5_0
flask-cors=3.0.10=pyhd3eb1b0_0
flask-socketio=5.3.1=py312hca03da5_0
gunicorn=21.2.0=py312h81bd7bf_1
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.3=py312hca03da5_0
libcxx=16.0.6=h4653b0c_0
libexpat=2.5.0=hb7217d7_1
libffi=3.4.4=hca03da5_0
libsqlite=3.45.2=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
markupsafe=2.1.3=py312h80987f9_0
ncurses=6.4=h313beb8_0
openssl=3.2.1=h0d3ecfb_1
packaging=23.2=py312hca03da5_0
pip=23.3.1=py312hca03da5_0
python=3.12.2=hdf0ec26_0_cpython
python-duckdb=0.10.1=py312h20a0b95_2
python-engineio=4.1.0=pyhd3eb1b0_0
python-socketio=5.3.0=pyhd3eb1b0_0
python_abi=3.12=4_cp312
readline=8.2=h1a28f6b_0
setuptools=68.2.2=py312hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h80987f9_0
tk=8.6.13=h5083fa2_1
tzdata=2024a=h04d1e81_0
werkzeug=2.3.8=py312hca03da5_0
wheel=0.41.2=py312hca03da5_0
xz=5.4.6=h80987f9_0
zlib=1.2.13=h53f4e23_5